refactor(placement): move the bench bridge write half into placement
`placement.py` picked a bridge and reported its headroom while
`docker_manager.py` created it and attached infrastructure to it, so a
change to bench placement touched two modules.

Move seven functions into `placement.py`: `subnet_base`,
`bench_network_spec`, `bench_network_index`, `ensure_bench_network`,
`ensure_bench_network_for`, `attach_infrastructure` and
`missing_infrastructure`, with the four constants only they read.
`missing_infrastructure` stays separate from `attach_infrastructure`:
a reconciler must report what it found apart from what it fixed.

`ensure_network` stays in `docker_manager`. It creates the app's own
shared `benchpress` network — one network, with no index, no slot and no
headroom — so it is not a placement decision.

Callers reach the moved names as `placement.<name>`, never a from-import,
so each has one patch target. `create_bench_container` imports
`placement` inside the function, as `start_bench_container` already does,
because `placement` imports `docker_manager` at module level.
