+++ benchpress/docker_manager.py
@@ -31,14 +31,6 @@ PREFLIGHT_IMAGE = "alpine"
 # The network every bench predating the bridge family sits on. It keeps receiving
 # MariaDB, Redis and Traefik as their primary network; it stops receiving benches.
 LEGACY_NETWORK = "benchpress"
-NETWORK_PREFIX = "benchpress-"
-BRIDGE_DEVICE_PREFIX = "bpbr"
-DEFAULT_SUBNET_BASE = "10.20"
-
-# Attached to every bench bridge so Docker's embedded DNS answers in both
-# directions. The control plane's own db and redis are deliberately absent: a
-# tenant bridge that reaches them is a breach, not a convenience.
-INFRASTRUCTURE_CONTAINERS = ("benchpress_traefik", "benchpress-mariadb", "benchpress-redis")
 
 
 def resolve_runtime(bench_doc) -> str | None:
@@ -177,105 +169,6 @@ def ensure_network(client: docker.DockerClient | None = None) -> None:
 		)
 
 
-def subnet_base() -> str:
-	return frappe.get_cached_doc("BenchPress Settings").bench_subnet_base or DEFAULT_SUBNET_BASE
-
-
-def bench_network_spec(index: int, base: str | None = None) -> dict:
-	"""Name, bridge device, subnet and gateway for one bench bridge.
-
-	`index * 16` in the third octet starts every /20 on its own boundary, so a later
-	per-node /16 scheme renumbers nothing.
-	"""
-	base = base or DEFAULT_SUBNET_BASE
-	octet = index * 16
-	return {
-		"name": f"{NETWORK_PREFIX}{index}",
-		"device": f"{BRIDGE_DEVICE_PREFIX}{index}",
-		"subnet": f"{base}.{octet}.0/20",
-		"gateway": f"{base}.{octet}.1",
-	}
-
-
-def bench_network_index(network: str) -> int | None:
-	"""The family index a network name encodes, or None if the name is not one of ours."""
-	if not network.startswith(NETWORK_PREFIX):
-		return None
-	suffix = network[len(NETWORK_PREFIX) :]
-	return int(suffix) if suffix.isdigit() else None
-
-
-def ensure_bench_network(index: int, client: docker.DockerClient | None = None) -> str:
-	"""Create bench bridge `index` if absent, attach infrastructure, return its name."""
-	client = client or get_client()
-	spec = bench_network_spec(index, subnet_base())
-	try:
-		client.networks.get(spec["name"])
-	except docker.errors.NotFound:
-		client.networks.create(
-			spec["name"],
-			driver="bridge",
-			ipam=docker.types.IPAMConfig(
-				pool_configs=[docker.types.IPAMPool(subnet=spec["subnet"], gateway=spec["gateway"])]
-			),
-			options={
-				# Honoured only at create. A network that already exists without it
-				# keeps its br-<hex> device forever, and phase 3's per-device
-				# proxy_arp sysctl has no stable path to write.
-				"com.docker.network.bridge.name": spec["device"],
-				"com.docker.network.bridge.enable_icc": "true",
-				"com.docker.network.bridge.enable_ip_masquerade": "true",
-			},
-		)
-	attach_infrastructure(spec["name"], client)
-	return spec["name"]
-
-
-def ensure_bench_network_for(network: str, client: docker.DockerClient | None = None) -> str:
-	"""Ensure the network a bench row names, whether legacy or one of the family."""
-	client = client or get_client()
-	if network == LEGACY_NETWORK:
-		ensure_network(client)
-		return network
-	index = bench_network_index(network)
-	if index is None:
-		frappe.throw(_("'{0}' is not a BenchPress bench network.").format(network))
-	return ensure_bench_network(index, client)
-
-
-def attach_infrastructure(network: str, client: docker.DockerClient | None = None) -> list[str]:
-	"""Connect the three infrastructure containers to `network`; returns those now on it.
-
-	Only INFRASTRUCTURE_CONTAINERS is ever attachable. The tuple is a security
-	boundary: an attachment reaches `Network.connect`, and putting the control
-	plane's own database on a network tenants share is a breach.
-	"""
-	client = client or get_client()
-	net = client.networks.get(network)
-	missing = missing_infrastructure(network, client)
-	attached = [name for name in INFRASTRUCTURE_CONTAINERS if name not in missing]
-	for name in missing:
-		try:
-			net.connect(name)
-			attached.append(name)
-		except docker.errors.NotFound:
-			continue  # a dev checkout has no Traefik
-		except docker.errors.APIError as e:
-			frappe.logger("benchpress").warning(f"could not attach {name} to {network}: {e}")
-	return attached
-
-
-def missing_infrastructure(network: str, client: docker.DockerClient | None = None) -> list[str]:
-	"""Infrastructure containers absent from `network`; empty when the bridge does not exist."""
-	client = client or get_client()
-	try:
-		net = client.networks.get(network)
-	except docker.errors.NotFound:
-		return []
-	present = {c.get("Name") for c in net.attrs.get("Containers", {}).values()}
-	return [name for name in INFRASTRUCTURE_CONTAINERS if name not in present]
-
-
 def build_lab_image(lab_doc, log_fn=None, no_cache: bool = False) -> str:
 	"""Build Docker image with bench + apps (site created at runtime against shared MariaDB).
 
@@ -335,9 +228,11 @@ def create_bench_container(bench_doc, lab_doc, network: str | None = None) -> st
 	`network` overrides the bench's recorded bridge, which is what a roll onto the next
 	bridge needs: the row still names the bridge Docker refused.
 	"""
+	from benchpress import placement  # placement imports this module
+
 	client = get_client()
 	network = network or getattr(bench_doc, "bridge_network", None) or LEGACY_NETWORK
-	ensure_bench_network_for(network, client)
+	placement.ensure_bench_network_for(network, client)
 
 	name = bench_doc.bench_name
